chore: add unit tests for Governing_Data_Controller and Search_Controller across different site types
